+++ GDEdit/GDEdit/Utilities/Objects/GeometryDash/LevelObjects/SpecialObjects/Pads/MagentaPad.cs
@@ -15,7 +15,7 @@ namespace GDEdit.Utilities.Objects.GeometryDash.LevelObjects.SpecialObjects.Pads
     public class MagentaPad : Pad
     {
         /// <summary>The object ID of the magenta pad.</summary>
-        public override short ObjectID => (short)(int)PadType.MagentaPad;
+        public override int ObjectID => (int)PadType.MagentaPad;
 
         /// <summary>Initializes a new instance of the <seealso cref="MagentaPad"/> class.</summary>
         public MagentaPad() : base() { }
