+++ GDEdit/GDEdit/Utilities/Objects/GeometryDash/LevelObjects/SpecialObjects/Portals/GamemodePortals/SpiderPortal.cs
@@ -14,7 +14,7 @@ namespace GDEdit.Utilities.Objects.GeometryDash.LevelObjects.SpecialObjects.Port
     public class SpiderPortal : GamemodePortal, IHasCheckedProperty
     {
         /// <summary>The object ID of the spider portal.</summary>
-        public override short ObjectID => (short)(int)PortalType.Spider;
+        public override int ObjectID => (int)PortalType.Spider;
         /// <summary>The gamemode the gamemode portal transforms the player into.</summary>
         public override Gamemode Gamemode => Gamemode.Spider;
