+++ GDEdit/GDEdit/Utilities/Objects/GeometryDash/LevelObjects/SpecialObjects/Portals/SpeedPortals/NormalSpeedPortal.cs
@@ -14,7 +14,7 @@ namespace GDEdit.Utilities.Objects.GeometryDash.LevelObjects.SpecialObjects.Port
     public class NormalSpeedPortal : SpeedPortal, IHasCheckedProperty
     {
         /// <summary>The object ID of the normal speed portal.</summary>
-        public override short ObjectID => (short)(int)PortalType.NormalSpeed;
+        public override int ObjectID => (int)PortalType.NormalSpeed;
 
         /// <summary>The speed this speed portal sets.</summary>
         public override Speed Speed => Speed.Normal;
