+++ GDEdit/GDEdit/Utilities/Objects/GeometryDash/LevelObjects/SpecialObjects/Portals/SpeedPortals/SlowSpeedPortal.cs
@@ -14,7 +14,7 @@ namespace GDEdit.Utilities.Objects.GeometryDash.LevelObjects.SpecialObjects.Port
     public class SlowSpeedPortal : SpeedPortal, IHasCheckedProperty
     {
         /// <summary>The object ID of the slow speed portal.</summary>
-        public override short ObjectID => (short)(int)PortalType.SlowSpeed;
+        public override int ObjectID => (int)PortalType.SlowSpeed;
 
         /// <summary>The speed this speed portal sets.</summary>
         public override Speed Speed => Speed.Slow;
