+++ GDEdit/GDEdit/Utilities/Objects/GeometryDash/LevelObjects/SpecialObjects/SpecialBlocks/DSpecialBlock.cs
@@ -11,7 +11,7 @@ namespace GDEdit.Utilities.Objects.GeometryDash.LevelObjects.SpecialObjects.Spec
     public class DSpecialBlock : SpecialBlock
     {
         /// <summary>The object ID of the D special block.</summary>
-        public override short ObjectID => (short)(int)SpecialBlockType.D;
+        public override int ObjectID => (int)SpecialBlockType.D;
 
         /// <summary>Initializes a new instance of the <seealso cref="DSpecialBlock"/> class.</summary>
         public DSpecialBlock() : base() { }
