+++ GDEdit/GDEdit/Utilities/Objects/GeometryDash/LevelObjects/SpecialObjects/SpecialBlocks/JSpecialBlock.cs
@@ -11,7 +11,7 @@ namespace GDEdit.Utilities.Objects.GeometryDash.LevelObjects.SpecialObjects.Spec
     public class JSpecialBlock : SpecialBlock
     {
         /// <summary>The object ID of the J special block.</summary>
-        public override short ObjectID => (short)(int)SpecialBlockType.J;
+        public override int ObjectID => (int)SpecialBlockType.J;
 
         /// <summary>Initializes a new instance of the <seealso cref="JSpecialBlock"/> class.</summary>
         public JSpecialBlock() : base() { }
