+++ GDEdit/GDEdit/Utilities/Objects/GeometryDash/LevelObjects/SpecialObjects/SpecialObject.cs
@@ -23,7 +23,7 @@ namespace GDEdit.Utilities.Objects.GeometryDash.LevelObjects.SpecialObjects
         /// <param name="objectID">The object ID of the rotating object.</param>
         /// <param name="x">The X location of the object.</param>
         /// <param name="y">The Y location of the object.</param>
-        public SpecialObject(short objectID, double x, double y)
+        public SpecialObject(int objectID, double x, double y)
             : base(objectID, x, y)
         {
             // Since this constructor must be only used by general object classes that may have different object IDs, assume the property is overriden
