+++ GDEdit/GDEdit/Utilities/Objects/GeometryDash/LevelObjects/Triggers/ColorTriggers/Color2ColorTrigger.cs
@@ -10,7 +10,7 @@ namespace GDEdit.Utilities.Objects.GeometryDash.LevelObjects.Triggers.ColorTrigg
     /// <summary>Represents a Color 2 Color trigger.</summary>
     public class Color2ColorTrigger : ColorTrigger
     {
-        public override short ObjectID => (short)(int)TriggerType.Color2;
+        public override int ObjectID => (int)TriggerType.Color2;
         
         /// <summary>The target Color ID of the trigger.</summary>
         public new short TargetColorID => 2;
