+++ GDEdit/GDEdit/Utilities/Objects/GeometryDash/LevelObjects/Triggers/ColorTriggers/Color4ColorTrigger.cs
@@ -10,7 +10,7 @@ namespace GDEdit.Utilities.Objects.GeometryDash.LevelObjects.Triggers.ColorTrigg
     /// <summary>Represents a Color 4 Color trigger.</summary>
     public class Color4ColorTrigger : ColorTrigger
     {
-        public override short ObjectID => (short)(int)TriggerType.Color4;
+        public override int ObjectID => (int)TriggerType.Color4;
         
         /// <summary>The target Color ID of the trigger.</summary>
         public new short TargetColorID => 4;
