+++ GDEdit/GDEdit/Utilities/Objects/GeometryDash/LevelObjects/Triggers/Interfaces/IHasTargetItemID.cs
@@ -13,6 +13,6 @@ namespace GDEdit.Utilities.Objects.GeometryDash.LevelObjects.Triggers.Interfaces
     {
         /// <summary>The target Item ID of the trigger.</summary>
         [ObjectStringMappable(ObjectParameter.ItemID)]
-        short TargetItemID { get; set; }
+        int TargetItemID { get; set; }
     }
 }
